docker/vectorization_markup/requirements.txt
torch==2.1.1
tokenizers==0.10.1
requests
tqdm
six
transformers==4.6.0
grpcio
grpcio-tools
numpy==1.21.4

